Move wavecontroller to model-package
